8 - декомпозировал страницу main, добавил сценарий вывода сообщения если не найдены офееры, удалил лишние страницы, список городов обернулв в memo
