--- project/src/pages/main-empty/main-empty.tsx
@@ -1,63 +1,0 @@
-import Header from '../../components/header/header';
-import { Helmet } from 'react-helmet-async';
-
-function MainEmpty(): JSX.Element
-{
-  return(
-    <div className="page page--gray page--main">
-      <Helmet><title>Шесть городов. Предложения не найдены.</title></Helmet>
-      <Header isNumberCitiesLogo={false} isLogoLinkActive isHaveHeaderNav/>
-      <main className="page__main page__main--index page__main--index-empty">
-        <h1 className="visually-hidden">Cities</h1>
-        <div className="tabs">
-          <section className="locations container">
-            <ul className="locations__list tabs__list">
-              <li className="locations__item">
-                <a className="locations__item-link tabs__item" href="/#">
-                  <span>Paris</span>
-                </a>
-              </li>
-              <li className="locations__item">
-                <a className="locations__item-link tabs__item" href="/#">
-                  <span>Cologne</span>
-                </a>
-              </li>
-              <li className="locations__item">
-                <a className="locations__item-link tabs__item" href="/#">
-                  <span>Brussels</span>
-                </a>
-              </li>
-              <li className="locations__item">
-                <a className="locations__item-link tabs__item" href="/#">
-                  <span>Amsterdam</span>
-                </a>
-              </li>
-              <li className="locations__item">
-                <a className="locations__item-link tabs__item" href="/#">
-                  <span>Hamburg</span>
-                </a>
-              </li>
-              <li className="locations__item">
-                <a className="locations__item-link tabs__item" href="/#">
-                  <span>Dusseldorf</span>
-                </a>
-              </li>
-            </ul>
-          </section>
-        </div>
-        <div className="cities">
-          <div className="cities__places-container cities__places-container--empty container">
-            <section className="cities__no-places">
-              <div className="cities__status-wrapper tabs__content">
-                <b className="cities__status">No places to stay available</b>
-                <p className="cities__status-description">We could not find any property available at the moment in Dusseldorf</p>
-              </div>
-            </section>
-            <div className="cities__right-section"></div>
-          </div>
-        </div>
-      </main>
-    </div>);
-}
-
-export default MainEmpty;
